services/ai/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
PyJWT==2.8.0
django-cors-headers==4.3.1

# ML/DL Dependencies (compatible with Python 3.12)
tensorflow==2.16.1
torch==2.2.0
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4

# NLP & Embeddings
transformers==4.37.0
sentence-transformers==2.3.0
openai==1.10.0

# Vector Database
chromadb==0.4.22
faiss-cpu==1.8.0

# API & HTTP
requests==2.31.0
httpx==0.26.0

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
